feat: 커밋 테이블 작성

+++ models/github_account.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, BigInteger, String, DateTime
 from sqlalchemy.orm import relationship
 from db import Base
 
 class GithubAccount(Base):
     __tablename__ = "github_account"
 
-    github_id = Column(Integer, primary_key=True, index=True)
+    github_id = Column(BigInteger, primary_key=True, index=True) # 깃허브에서 제공하는 고유 id
     github_login = Column(String(255), nullable=False)
     github_token = Column(String(255), nullable=True)
     last_synced_at = Column(DateTime)
